# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
...
			"EmailTemplateDoc": {
				"properties": {
					"name": {
						"type": "string"
					},
					"subject": {
						"type": "string"
					},
					"html": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"variables": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"subject",
					"html"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SuccessResponse_EmailTemplateDoc_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/EmailTemplateDoc"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"ErrorResponse": {
				"properties": {
					"timestamp": {
						"type": "string"
					},
					"error": {
						"properties": {
							"details": {},
							"code": {
								"type": "string"
							},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"error",
					"message",
					"success"
				],
				"type": "object"
			},
			"ApiResponse_EmailTemplateDoc_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_EmailTemplateDoc_"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
			"SuccessResponse_EmailTemplateDoc-Array_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/EmailTemplateDoc"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"ApiResponse_EmailTemplateDoc-Array_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_EmailTemplateDoc-Array_"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
			"SuccessResponse_null_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"ApiResponse_null_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_null_"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
			"Partial_EmailTemplateDoc_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"subject": {
						"type": "string"
					},
					"html": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"variables": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"SuccessResponse_any_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
...
			"NewsletterSubscriber": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"email": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SuccessResponse_NewsletterSubscriber_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/NewsletterSubscriber"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"ApiResponse_NewsletterSubscriber_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_NewsletterSubscriber_"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
...
			"ApiResponse_any_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_any_"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
...
			"UploadedFileDoc": {
				"properties": {
					"filename": {
						"type": "string"
					},
					"mimetype": {
						"type": "string"
					},
					"extension": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"filename",
					"mimetype",
					"extension",
					"path",
					"size",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SuccessResponse__files-UploadedFileDoc-Array--total-number__": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"properties": {
							"total": {
								"type": "number",
								"format": "double"
							},
							"files": {
								"items": {
									"$ref": "#/components/schemas/UploadedFileDoc"
								},
								"type": "array"
							}
						},
						"required": [
							"total",
							"files"
						],
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"ApiResponse__files-UploadedFileDoc-Array--total-number__": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse__files-UploadedFileDoc-Array--total-number__"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "app",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/v1/email-templates": {
			"post": {
				"operationId": "CreateTemplate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_EmailTemplateDoc_"
								}
							}
						}
					}
				},
				"description": "Create a new template",
				"tags": [
					"email-templates"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmailTemplateDoc"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetTemplates",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_EmailTemplateDoc-Array_"
								}
							}
						}
					}
				},
				"description": "Get all templates",
				"tags": [
					"email-templates"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/email-templates/{id}": {
			"post": {
				"operationId": "SendEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"email-templates"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_EmailTemplateDoc_"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateTemplate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_EmailTemplateDoc_"
								}
							}
						}
					}
				},
				"description": "Update a template by ID",
				"tags": [
					"email-templates"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_EmailTemplateDoc_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTemplate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Delete a template by ID",
				"tags": [
					"email-templates"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/newsletter/email/{email}": {
			"get": {
				"operationId": "GetSubscriberByEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_NewsletterSubscriber_"
								}
							}
						}
					}
				},
				"tags": [
					"newsletter"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/files": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"files"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetFiles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__files-UploadedFileDoc-Array--total-number__"
								}
							}
						}
					}
				},
				"tags": [
					"files"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"createdAt",
								"filename",
								"mimetype",
								"extension",
								"path",
								"size"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/files/{id}": {
			"delete": {
				"operationId": "DeleteFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"files"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}